Sonarqube

+++ anchor-plugin-image-feature/src/main/java/org/anchoranalysis/plugin/image/feature/bean/score/DifferenceHelper.java
@@ -37,7 +37,7 @@ class DifferenceHelper {
             double widthGreaterThan,
             double widthLessThan,
             int minDifference) {
-        double diff = (double) (valueFirst - valueSecond - minDifference);
+        double diff = valueFirst - valueSecond - minDifference;
 
         if (diff < (-1 * widthLessThan)) {
             return 0.0;
